refactor: move functions into separated files
